refactor(symbol_analysis): split graph.rs into mod/walker/tests submodule
`graph.rs` was 1744 LOC and tripped the "Reject .rs files over 1000
LOC" CI gate. Also fixes the `Documentation` gate which was failing
on an unresolved rustdoc intra-doc link.

Restructure: `symbol_analysis/graph.rs` → `symbol_analysis/graph/`
with three files, each well under the 1000-LOC limit:

- `mod.rs` (698) — public types (`GraphConfig`, `GraphNode`,
  `EdgeDirection`, `GraphEdge`, `BackrefGraph`), the
  `BackrefGraph::build*` entry points, `to_dot()` rendering, and the
  small dot-helpers (`node_color`, `node_width`, `format_*_label`,
  `sanitize_id`, `sanitize_filename`, `escape_dot_label`).
- `walker.rs` (361) — internal walker (`CappedReferencer`,
  `rank_and_cap_referencers`, `push_edge_dedup`, `walk_forward`,
  `CalleeCandidate`, `format_callee_label`) plus the public
  `rank_callees_dual` / `CalleeRanked` API, re-exported from
  `mod.rs` so the existing path
  `fbuild_core::symbol_analysis::graph::rank_callees_dual` keeps
  resolving for `fbuild_build::symbol_analyzer`.
- `tests.rs` (715) — the entire `#[cfg(test)] mod tests` body.

Also fixes the rustdoc broken-link warning that failed the
`Documentation` job: `[`build`]` → `[`Self::build`]` so the
intra-doc resolver finds it (`-D rustdoc::broken_intra_doc_links`
under `-D warnings`).

No public API change. Content-preserving refactor; only visibility
adjustments (`pub(super)` where the parent calls into the new
walker module).

Verified clean: `check`, `clippy --all-targets -- -D warnings`,
`test -p fbuild-core --lib symbol_analysis` (66 passed including
all 20 `graph::tests::*`), `fmt --check`, and
`RUSTDOCFLAGS="-D warnings" cargo doc -p fbuild-core --no-deps`.
